move db logic to db folder
