+++ src-tauri/src/commands/category.rs
@@ -1,62 +1,29 @@
 use tauri::{command, State};
 
 use crate::{
-    errors::{CommandError, CommandResult},
+    db,
+    errors::CommandResult,
     schema::{AppState, Category},
 };
 
 #[command]
-#[allow(clippy::needless_pass_by_value)]
 pub fn get_categories(state: State<AppState>) -> CommandResult<Vec<Category>> {
-    let db = state.db.conn.lock().unwrap();
-    let mut categories = Vec::new();
-
-    let mut stmt = db.prepare("SELECT * FROM categories")?;
-    dbg!(&stmt);
-    let rows = stmt.query_map([], |row| {
-        Ok(Category {
-            category_id: row.get(0)?,
-            name: row.get(1)?,
-            description: row.get(2).unwrap_or(String::new()),
-        })
-    })?;
-    for category in rows {
-        categories.push(category?);
-    }
-    Ok(categories)
+    let conn = state.db.conn.lock().unwrap();
+    db::get_categories(&conn)
 }
 
 #[command]
-#[allow(clippy::needless_pass_by_value)]
 pub fn create_category(
     name: String,
     description: String,
     state: State<AppState>,
 ) -> CommandResult<()> {
-    let db = state.db.conn.lock().unwrap();
-    let mut stmt = db.prepare("INSERT INTO categories (name, description) VALUES ( ?, ? )")?;
-
-    if let Err(e) = stmt.execute([name, description]) {
-        return Err(CommandError::RusqliteError(e));
-    };
-
-    Ok(())
+    let conn = state.db.conn.lock().unwrap();
+    db::create_category(&conn, name, description)
 }
 
 #[command]
-#[allow(clippy::needless_pass_by_value)]
 pub fn delete_category(id: usize, state: State<AppState>) -> CommandResult<()> {
-    let db = state.db.conn.lock().unwrap();
-
-    // let mut stmt = db.prepare("DELETE FROM items_to_categories WHERE category_id = ( SELECT id FROM categories WHERE id = ? )")?;
-    // if let Err(e) = stmt.execute([id]) {
-    //     return Err(CommandError::RusqliteError(e));
-    // };
-
-    let mut stmt = db.prepare("DELETE FROM categories WHERE category_id = ?")?;
-    if let Err(e) = stmt.execute([id]) {
-        return Err(CommandError::RusqliteError(e));
-    };
-
-    Ok(())
+    let conn = state.db.conn.lock().unwrap();
+    db::delete_category(&conn, id)
 }
