+++ src-tauri/src/db.rs
@@ -1,3 +1,7 @@
+mod category;
 mod item;
+mod rating;
 
+pub use category::*;
 pub use item::*;
+pub use rating::*;
