+++ src-tauri/src/db/item.rs
@@ -24,6 +24,35 @@ pub fn get_item(conn: &MutexGuard<Connection>, id: usize) -> CommandResult<Item>
     Ok(item)
 }
 
+pub fn get_items(conn: &MutexGuard<Connection>) -> CommandResult<Vec<Item>> {
+    let mut items = Vec::new();
+    let mut stmt = conn.prepare("SELECT * FROM items")?;
+    let rows = stmt.query_map([], |row| {
+        Ok(Item {
+            item_id: row.get(0)?,
+            name: row.get(1)?,
+            description: row.get(2).unwrap_or(String::new()),
+            comments: row.get(3).unwrap_or(String::new()),
+        })
+    })?;
+    for item in rows {
+        items.push(item?);
+    }
+    Ok(items)
+}
+
+pub fn update_item(conn: &MutexGuard<Connection>, item: Item) -> CommandResult<usize> {
+    let mut stmt =
+        conn.prepare("UPDATE items SET name = ?, description = ?, comments = ? WHERE item_id = ?")?;
+    let id = stmt.execute([
+        item.name,
+        item.description,
+        item.comments,
+        item.item_id.to_string(),
+    ])?;
+    Ok(id)
+}
+
 pub fn create_item(
     conn: &MutexGuard<Connection>,
     item: Item,
